requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.6
openai==1.65.5
python-multipart
python-jose[cryptography]
passlib[bcrypt]
asyncpg==0.29.0
email-validator==2.1.0.post1
pydantic[email]
greenlet==3.0.3
httpx==0.28.1
